Make git try work on Windows.
This mirrors the git-cl changes by first moving git-try
to git_try.py (which git cl already had), then copying
git-cl to git-try and updating all references to git-cl
and git_cl.py to the corresponding git try references.

The main issue is that in the MINGW/Msysgit case we want
to use the Python in depot_tools since that is 100%
compatible with the Command Prompt.  Hence we need
to call it directly and then pass it the name of the
git_try.py script.

Note that 'git-try' will not (ever) work on Windows,
and users should only ever run 'git try'.  Hence the
need to modify the output of the command to avoid
further confusion.
Review URL: http://codereview.chromium.org/7114002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@89031 0039d316-1c4b-4281-b951-d872f2087c98