Add reviewers to optional outputs for changes
This adds a comma-separated list of reviewers to the optional outputs
for changes. This can be added to the output using the custom formatter
option.  E.g. --output-format-changes="{url} {title} <{reviewers}>"

Also, for webkit changes, use CommitterList to map reviewer names to
emails so that they are similar format to other changes.


Review URL: https://chromiumcodereview.appspot.com/12446008

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@187351 0039d316-1c4b-4281-b951-d872f2087c98