Always convert flag value to string in set_flag
The callee needs it to be a string so that it can be made
part of HTTP request. However, in code using Rietveld
it's useful to use non-string types for flags where
Rietveld actually doesn't return strings (say booleans).

One example is the commit flag, which is a boolean.

BUG=291335

Review URL: https://chromiumcodereview.appspot.com/24846002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@225748 0039d316-1c4b-4281-b951-d872f2087c98