Indicate when depot_tools can't be updated on Windows.
When depot_tools fail to update right now all you get is
    Cannot rebase: You have unstaged changes.
    Please commit or stash them.
As that is typically in a call to gclient it's far from obvious what
could not be updated so add a line saying
    Failed to update depot_tools.
to that. Linux and Mac already have a similar error message in their
script file.

BUG=
R=szager@chromium.org

Review URL: https://codereview.chromium.org/23619058

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@226002 0039d316-1c4b-4281-b951-d872f2087c98