Reapply 231888 "Fixes for python 2.7.5 support." with fixes.
Tested to work now with the repro case of bug 313552.

"/c" was interpreted by cygwin's bash as a path and was replaced with c:/. :(
Using //c seems to be passed through properly.
There is probably other escaping problem but this should be good enough at
this stage.

Ref: http://www.mingw.org/wiki/Posix_path_conversion

R=iannucci@chromium.org
BUG=241769
BUG=313552

Review URL: https://codereview.chromium.org/50143013

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@232184 0039d316-1c4b-4281-b951-d872f2087c98