Support Gerrit snapshot versions
These may be named git-snapshot-VERSION.war, where VERSION is the long
format defined in git-describe(1). In particular, this may include a
number of commits after the given tag name; treat more commits since
the tag as a higher version number. This is arguably not always
strictly correct due to non-linear history, but the maintainers expect
the total number of such snapshot versions to be low, and can remove
old snapshot versions that don't sort correctly.

Slightly de-obfuscate some of the variable names and logic while we're
at it.

R=szager@chromium.org
BUG=

Review URL: https://codereview.chromium.org/164023005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@251888 0039d316-1c4b-4281-b951-d872f2087c98