Fix GitCheckout to be sane and use git-cache
The commit queue uses checkout.GitCheckout to manage the
local git repository. It resets it to a clean state,
commits patches locally, and pushes those changes to the
golden-source-of-truth repo.

This CL fixes the prepare() method to use the new git-cache
tool, and in the process makes it more consistent, more
readable, and more correct. It also makes minor changes to
other methods to bring them in line with the git-cache
approach.

R=iannucci@chromium.org, rmistry@chromium.org
BUG=261619

Review URL: https://codereview.chromium.org/175873003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@256681 0039d316-1c4b-4281-b951-d872f2087c98