Do not fetch the CL description when using git cl status --fast.
This causes a network I/O, which is not fast. Also this is a non-starter in the
case of no network connectivity, causing a crash instead of completing
successfully. Until Air Canada provides on-flight internet, this is a
non-starter. #firstworldproblem.

Also fix git cl presubmit in case of lack of network connectivity.

R=stip@chromium.org
BUG=

Review URL: https://codereview.chromium.org/418293004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@285911 0039d316-1c4b-4281-b951-d872f2087c98