Allow to skip revision check in roll-dep.
Without this change, roll-dep requires dependencies to be
checked out and up to date in order to roll a newer revision.

This change allows to use roll-dep with less checking and
makes the script user responsible to verify the correctness
of the deps path and git hash.

This eases writing a deps roller that updates multiple deps
including a variety of os deps, which might not be checked
out on the host of the deps roller.

BUG=438705, 437462
R=jochen@chromium.org

Review URL: https://codereview.chromium.org/801643004

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@293444 0039d316-1c4b-4281-b951-d872f2087c98