Update toolchain packaging to support VC++ preview toolsets
This change adds a --override option to the VC++ packaging script so
that an arbitrary bin/include/lib containing directory can be used to
package up those components of VC++. See this post for details:

https://blogs.msdn.microsoft.com/vcblog/2016/02/16/try-out-the-latest-c-compiler-toolset-without-waiting-for-the-next-update-of-visual-studio/

Typical usage is like this:

  package_from_installed.py 2015 --override <projdir>\packages\VisualCppTools.14.0.24109-Pre\lib\native

In order to make this work the path tuple support was fixed so that the
local relative path was propagated to the destination path, instead of
just the file name.

At the same time, the code to package up the UCRT installers was deleted
because it isn't needed, and windows.winmd is filtered out because it
makes consistent package creation impossible (it is different on every
machine).

While removing the UCRT installers I discovered nested if VS_VERSION
checks. Oops. Fixed.

BUG=440500

Review-Url: https://codereview.chromium.org/1967653002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@300527 0039d316-1c4b-4281-b951-d872f2087c98