Make "vs_toolchain.py update" work on case-sensitive file systems
Filenames must have their case preserved so the files can be opened,
but they need to be lowercased when hashed and sorted, to match the
expected hash.

In other words, this patch moves the lower-casing from GetFileList
(except to sort the list) to where filenames are fed into the hash.

BUG=495204

Review-Url: https://codereview.chromium.org/2039563002

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/depot_tools@300696 0039d316-1c4b-4281-b951-d872f2087c98