Gerrit git cl: allow overriding GERRIT_SQUASH_UPLOADS in git config.
Previously, if GERRIT_SQUASH_UPLOADS was found in codereview.settings,
a local repository value of gerrit.squash-uploads would be set.

This CL changes that, so that gerrit.squash-uploads property is written
with GERRIT_SQUASH_UPLOADS value ONLY IF the property wasn't set before.

This allows individual users to choose their own way with
  git config [--global] -bool gerrit.squash-uploads (true|false)

BUG=611892
R=andybons@chromium.org,sergiyb@chromium.org

Review-Url: https://codereview.chromium.org/2070013002