Remove force parameter from bot_update.ensure_checkout.
This is the simplest change I can make here. The bot_update.py script already
ignores the value of force (and sets it to True in all cases). This is an
expectations-only change.

I'll refactor the content of the bot_update.py script to account for this
removed option in a followup CL as well.

R=agable@chromium.org, estaab@chromium.org, hinoka@chromium.org, martiniss@chromium.org
BUG=643885

Review-Url: https://codereview.chromium.org/2316323002