[cipd] Allow packaging whole root.
Currently, the "cipd" recipe module allows a package to be created by
adding directories to a PackageDefinition via "add_dir". However, there
is no current way to add the entirety of a root directory to a package.
The PackageDefinition will generate a relative path of "''", which CIPD
will reject as empty.

Make it so that if the user supplies the root directory as the directory
to add to the package, it properly includes it as ".".

BUG=None
TEST=local
  - Ran recipe with this modification, seems to work

R=iannucci@chromium.org, nodir@chromium.org, vadimsh@chromium.org

Change-Id: Ib5c72038cf153776808f084db835c60f61fd9044
Reviewed-on: https://chromium-review.googlesource.com/510309
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Daniel Jacques <dnj@chromium.org>