Use rstrip when getting footer tags
A quick bugfix was made in cc29098042 that removed trailing empty lines
when getting tags out of the git log.  This preserved the behavior that
tags with trailing whitespace would retain said whitespace.

This change instead simplifies the logic that removes trailing
whitespace with the side effect that whitespace at the end of the last
tag is also stripped.  The unittests are adjusted accordingly

Bug: 1130601
Change-Id: I26d39736179b36ac2573de1ca003a5bc09f30a28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/2423050
Auto-Submit: Shahbaz Youssefi <syoussefi@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>